[Bronze II] Title: 숫자, Time: 12 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.cc
@@ -17,6 +17,6 @@ int main() {
 
 	cout << b - a - 1 << '\n';
 
-	for (int i = a + 1; i < b; i++)
+	for (long long i = a + 1; i < b; i++)
 		cout << i << ' ';
 }
